Request protocol, - HTTPRequest to conform to Request protocol - change HTTPRequest from class to struct
