+++ Sources/Requests/HTTPRequestMethod.swift
@@ -1,0 +1,3 @@
+public enum HTTPRequestMethod: String {
+  case Get, Post, Put, Delete, Patch, Options, Head, Invalid
+}
