+++ Tests/ControllersTests/FormControllerTest.swift
@@ -6,7 +6,7 @@ import Util
 class FormControllerTest: XCTestCase {
   func testItReturnsEmptyOnInitialGet() {
     let rawRequest = "GET /form HTTP/1.1\r\nHost:\r\nConnection:Keep-Alive\r\nUser-Agent:chrome\r\nAccept-Encoding:gzip,deflate"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
 
     let contents: ControllerData = ControllerData(["form": Array("".utf8)])
 
@@ -18,8 +18,8 @@ class FormControllerTest: XCTestCase {
   func testItSavesDataOnPostReturnsOnGet() {
     let rawPostRequest = "POST /form HTTP/1.1\r\nHost:\r\nConnection:Keep-Alive\r\nUser-Agent:chrome\r\nAccept-Encoding:gzip,deflate\r\ndata=fatcat"
     let rawGetRequest = "GET /form HTTP/1.1\r\nHost:\r\nConnection:Keep-Alive\r\nUser-Agent:chrome\r\nAccept-Encoding:gzip,deflate"
-    let postRequest = Request(for: rawPostRequest)
-    let getRequest = Request(for: rawGetRequest)
+    let postRequest = HTTPRequest(for: rawPostRequest)
+    let getRequest = HTTPRequest(for: rawGetRequest)
 
     let contents: ControllerData = ControllerData(["form": Array("".utf8)])
 
@@ -35,9 +35,9 @@ class FormControllerTest: XCTestCase {
     let rawPostRequest = "POST /form HTTP/1.1\r\nHost:\r\nConnection:Keep-Alive\r\nUser-Agent:chrome\r\nAccept-Encoding:gzip,deflate\r\ndata=fatcat"
     let rawPutRequest = "PUT /form HTTP/1.1\r\nHost:\r\nConnection:Keep-Alive\r\nUser-Agent:chrome\r\nAccept-Encoding:gzip,deflate\r\ndata=hamilton"
     let rawGetRequest = "GET /form HTTP/1.1\r\nHost:\r\nConnection:Keep-Alive\r\nUser-Agent:chrome\r\nAccept-Encoding:gzip,deflate"
-    let postRequest = Request(for: rawPostRequest)
-    let putRequest = Request(for: rawPutRequest)
-    let getRequest = Request(for: rawGetRequest)
+    let postRequest = HTTPRequest(for: rawPostRequest)
+    let putRequest = HTTPRequest(for: rawPutRequest)
+    let getRequest = HTTPRequest(for: rawGetRequest)
 
     let contents: ControllerData = ControllerData(["form": Array("".utf8)])
 
@@ -54,9 +54,9 @@ class FormControllerTest: XCTestCase {
     let rawPostRequest = "POST /form HTTP/1.1\r\nHost:\r\nConnection:Keep-Alive\r\nUser-Agent:chrome\r\nAccept-Encoding:gzip,deflate\r\ndata=fatcat"
     let rawDeleteRequest = "DELETE /form HTTP/1.1\r\nHost:\r\nConnection:Keep-Alive\r\nUser-Agent:chrome\r\nAccept-Encoding:gzip,deflate"
     let rawGetRequest = "GET /form HTTP/1.1\r\nHost:\r\nConnection:Keep-Alive\r\nUser-Agent:chrome\r\nAccept-Encoding:gzip,deflate"
-    let postRequest = Request(for: rawPostRequest)
-    let deleteRequest = Request(for: rawDeleteRequest)
-    let getRequest = Request(for: rawGetRequest)
+    let postRequest = HTTPRequest(for: rawPostRequest)
+    let deleteRequest = HTTPRequest(for: rawDeleteRequest)
+    let getRequest = HTTPRequest(for: rawGetRequest)
 
     let contents: ControllerData = ControllerData(["form": Array("".utf8)])
