+++ Tests/ControllersTests/MethodOptionsControllerTest.swift
@@ -5,7 +5,7 @@ import XCTest
 class MethodOptionsControllerTest: XCTestCase {
   func testItCanRespondWithStatus200OnMethodOptionsRequest() {
     let rawRequest = "OPTIONS /method_options HTTP/1.1\r\n Host: localhost:5000\r\n Connection: Keep-Alive\r\n User-Agent: Apache-HttpClient/4.3.5 (java 1.5)\r\n Accept-Encoding: gzip,deflate"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
 
     let response = MethodOptionsController().process(request)
 
@@ -14,7 +14,7 @@ class MethodOptionsControllerTest: XCTestCase {
 
   func testItCanRespondWithCorrectAllowHeaders() {
     let rawRequest = "OPTIONS /method_options HTTP/1.1\r\n Host: localhost:5000\r\n Connection: Keep-Alive\r\n User-Agent: Apache-HttpClient/4.3.5 (java 1.5)\r\n Accept-Encoding: gzip,deflate"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
 
     let response = MethodOptionsController().process(request)
 
@@ -25,7 +25,7 @@ class MethodOptionsControllerTest: XCTestCase {
 
   func testItCanRespondWithStatus200MethodOptions2Request() {
     let rawRequest = "OPTIONS /method_options2 HTTP/1.1\r\n Host: localhost:5000\r\n Connection: Keep-Alive\r\n User-Agent: Apache-HttpClient/4.3.5 (java 1.5)\r\n Accept-Encoding: gzip,deflate"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
 
     let response = MethodOptionsController().process(request)
 
@@ -34,7 +34,7 @@ class MethodOptionsControllerTest: XCTestCase {
 
   func testItCanRespondWithCorrectHeadersMethodOptions2Request() {
     let rawRequest = "OPTIONS /method_options2 HTTP/1.1\r\n Host: localhost:5000\r\n Connection: Keep-Alive\r\n User-Agent: Apache-HttpClient/4.3.5 (java 1.5)\r\n Accept-Encoding: gzip,deflate"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
 
     let response = MethodOptionsController().process(request)
