+++ Tests/ControllersTests/ParametersControllerTest.swift
@@ -5,7 +5,7 @@ import XCTest
 class ParametersControllerTest: XCTestCase {
   func testItCanHandleNoParams() {
     let rawRequest = "GET /parameters HTTP/1.1\r\nHost: localhost:5000\r\nConnection: Keep-Alive\r\nUser-Agent: Apache-HttpClient/4.3.5 (java 1.5)\r\nAccept-Encoding: gzip,deflate"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
 
     let res = ParametersController().process(request)
 
@@ -14,7 +14,7 @@ class ParametersControllerTest: XCTestCase {
 
   func testItCanDecodeSimpleParams() {
     let rawRequest = "GET /parameters?variable_1=Operators HTTP/1.1\r\nHost: localhost:5000\r\nConnection: Keep-Alive\r\nUser-Agent: Apache-HttpClient/4.3.5 (java 1.5)\r\nAccept-Encoding: gzip,deflate"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
 
     let expected = Array("\n\nvariable_1 = Operators".utf8)
 
@@ -25,7 +25,7 @@ class ParametersControllerTest: XCTestCase {
 
   func testItCanDecodeMoreComplexParams() {
     let rawRequest = "GET /parameters?variable_1=Operators%20%3C%2C%20%3E%2C%20%3D HTTP/1.1\r\nHost: localhost:5000\r\nConnection: Keep-Alive\r\nUser-Agent: Apache-HttpClient/4.3.5 (java 1.5)\r\nAccept-Encoding: gzip,deflate"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
 
     let expected = Array("\n\nvariable_1 = Operators <, >, =".utf8)
 
@@ -36,7 +36,7 @@ class ParametersControllerTest: XCTestCase {
 
   func testItCanDecodeVeryComplexParams() {
     let rawRequest = "GET /parameters?variable_1=Operators%20%3C%2C%20%3E%2C%20%3D%2C%20!%3D%3B%20%2B%2C%20-%2C%20*%2C%20%26%2C%20%40%2C%20%23%2C%20%24%2C%20%5B%2C%20%5D%3A%20%22is%20that%20all%22%3F HTTP/1.1\r\nHost: localhost:5000\r\nConnection: Keep-Alive\r\nUser-Agent: Apache-HttpClient/4.3.5 (java 1.5)\r\nAccept-Encoding: gzip,deflate"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
 
     let expected = Array("\n\nvariable_1 = Operators <, >, =, !=; +, -, *, &, @, #, $, [, ]: \"is that all\"?".utf8)
 
@@ -47,7 +47,7 @@ class ParametersControllerTest: XCTestCase {
 
   func testItCanDecodeMultipleParams() {
     let rawRequest = "GET /parameters?variable_1=Operators%20%3C%2C%20%3E%2C%20%3D%2C%20!%3D%3B%20%2B%2C%20-%2C%20*%2C%20%26%2C%20%40%2C%20%23%2C%20%24%2C%20%5B%2C%20%5D%3A%20%22is%20that%20all%22%3F&variable_2=stuff HTTP/1.1\r\nHost: localhost:5000\r\nConnection: Keep-Alive\r\nUser-Agent: Apache-HttpClient/4.3.5 (java 1.5)\r\nAccept-Encoding: gzip,deflate"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
 
     let expected = Array("\n\nvariable_1 = Operators <, >, =, !=; +, -, *, &, @, #, $, [, ]: \"is that all\"?\nvariable_2 = stuff".utf8)
 
@@ -58,7 +58,7 @@ class ParametersControllerTest: XCTestCase {
 
   func testItDoesNotTryToDecodeInvalidParams() {
     let rawRequest = "GET /parameters?variable_1=Operators%229032%kansjd%029932030%owow HTTP/1.1\r\nHost: localhost:5000\r\nConnection: Keep-Alive\r\nUser-Agent: Apache-HttpClient/4.3.5 (java 1.5)\r\nAccept-Encoding: gzip,deflate"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
 
     let expected = Array("\n\nvariable_1 = Operators%229032%kansjd%029932030%owow".utf8)
