+++ Tests/ControllersTests/PartialContentsController.swift
@@ -5,7 +5,7 @@ import XCTest
 class PartialContentsControllerTest: XCTestCase {
   func testItCanRespondWithA206WithPartialContent() {
     let rawRequest = "GET /partial_content.txt HTTP/1.1\r\n Host: localhost:5000\r\n Connection: Keep-Alive\r\n User-Agent: Apache-HttpClient/4.3.5 (java 1.5)\r\n Accept-Encoding: gzip,deflate\r\nRange:bytes=4-"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
     let range = request.headers["range"]!
 
     let content = "This is a file that contains text to read part of in order to fulfill a 206."
@@ -17,7 +17,7 @@ class PartialContentsControllerTest: XCTestCase {
 
   func testItCanRespondWithPartialContentsGivenRangeStart() {
     let rawRequest = "GET /partial_content.txt HTTP/1.1\r\n Host: localhost:5000\r\n Connection: Keep-Alive\r\nRange:bytes=4-\r\n User-Agent: Apache-HttpClient/4.3.5 (java 1.5)\r\n Accept-Encoding: gzip,deflate"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
     let range = request.headers["range"]!
 
     let content = "This is a file that contains text to read part of in order to fulfill a 206."
@@ -31,7 +31,7 @@ class PartialContentsControllerTest: XCTestCase {
 
   func testItCanRespondWithPartialContentsGivenRangeEnd() {
     let rawRequest = "GET /partial_content.txt HTTP/1.1\r\n Host: localhost:5000\r\n Connection: Keep-Alive\r\nRange:bytes=-6\r\n User-Agent: Apache-HttpClient/4.3.5 (java 1.5)\r\n Accept-Encoding: gzip,deflate"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
     let range = request.headers["range"]!
 
     let content = "This is a file that contains text to read part of in order to fulfill a 206."
@@ -45,7 +45,7 @@ class PartialContentsControllerTest: XCTestCase {
 
   func testItCanRespondWithPartialContentsGivenRangeStartAndEnd() {
     let rawRequest = "GET /partial_content.txt HTTP/1.1\r\n Host: localhost:5000\r\n Connection: Keep-Alive\r\nRange:bytes=0-4\r\n User-Agent: Apache-HttpClient/4.3.5 (java 1.5)\r\n Accept-Encoding: gzip,deflate"
-    let request = Request(for: rawRequest)
+    let request = HTTPRequest(for: rawRequest)
     let range = request.headers["range"]!
 
     let content = "This is a file that contains text to read part of in order to fulfill a 206."
